+++ src/routers/app.js
@@ -5,7 +5,6 @@ module.exports = function (app) {
     const router = express.Router();
     const passport = require('passport');
     const LocalStrategy = require('passport-local').Strategy;
-    const schedule = require('node-schedule');
 
     const Controller = app.controllers.app;
     const User = app.models.user;
@@ -13,6 +12,15 @@ module.exports = function (app) {
 
     app.use(passport.initialize());
     app.use(passport.session());
+
+    passport.serializeUser((user, done) => {
+        done(null, user.id);
+    });
+    passport.deserializeUser((id, done) => {
+        User.findById(id, (err, user) => {
+            done(err, user);
+        });
+    });
 
     passport.use(new LocalStrategy({
         usernameField: 'email',
@@ -34,15 +42,69 @@ module.exports = function (app) {
         });
     }));
 
-    passport.serializeUser((user, done) => {
-        done(null, user.id);
-    });
-    passport.deserializeUser((id, done) => {
-        User.findById(id, (err, user) => {
-            done(err, user);
-        });
-    });
+    // Handle "stay connected" feature, if enabled, in passport auth.
+    let sessionKeeperCookieOptions = Controller.$getSessionKeeperCookieOptions(app.get('config'));
+    if (sessionKeeperCookieOptions !== false) {
 
+        /*
+         * Initialise the remember me strategy, giving it cookie options from config and 2 callbacks:
+         *  1. the verify callback, which check if a user exists base on the remember me cookie value
+         *  2. a callback to generate a new cookie for remembering the user, if verify callback succeed to indentify a
+         *     user: it's recommended to regenerate a fresh token once a previous one was successfully used for auth
+         */
+        const RememberMeStrategy = require('lib/passport-remember-me').Strategy;
+        let {name:cookieName, ...cookieOptions} = sessionKeeperCookieOptions;
+
+        passport.use(new RememberMeStrategy({
+                key: cookieName,
+                cookie: cookieOptions
+            },
+            async (req, cookieValue, done) => {
+                // The cookie contain a random uniq ID and a random uniq token, separated by a comma (,)
+                // The ID must match a user, and the token must match a hash associated to this ID.
+                let [id, token] = cookieValue.split(',');
+                // Search for a user matching ID and token
+                let user = await User.findBySessionKeeperToken(id, token);
+                if (user) {
+                    // A user was found matching the cookie, it'll be automatically logged in, after some cleaning
+                    // Remove the session keeper id used to identify a user. It's secure to consider them as
+                    // only one time use
+                    user.removeSessionKeeperTokens(id);
+                    // Also remove all other tokens for the current "computer"
+                    let {userAgent, remoteIp} = Controller.$getUserComputerIdentificationData(req);
+                    user.clearSessionKeeperTokens(userAgent, remoteIp);
+                    // Save all that in database
+                    await user.save();
+                    // Return the authenticated user to passport for login finalize
+                    return done(null, user);
+                }
+                // No user found, return it to the Strategy (this is not a fail, just we don't had a chance to
+                // remember a user and autologin it, so a standard auth will now occurs)
+                return done(null, false)
+            },
+            async (req, user, done) => {
+                // Generate a new token for session keeping for the user, in order to keep autoloading him next time.
+                // This mecanisme (delete the used token and generate a new one) is to improve security. It also
+                // reset the session keeper (its cookie) life.
+                let config = req.app.get('config');
+                let {userAgent, remoteIp} = Controller.$getUserComputerIdentificationData(req);
+                let {id, token} = user.createSessionKeeperToken(userAgent, remoteIp, config.get('session:stayConnected:expiration'));
+                await user.save();
+                // Return the new session keeper cookie value to the passport strategy, which will save the cookie
+                let cookieValue = id + "," + token;
+                return done(null, cookieValue);
+            }
+        ));
+
+        // Declare the middleware for the remember me passport strategy
+        app.use(passport.authenticate('remember-me'));
+    }
+
+    /* Generic middleware for all the app routes to set locals for templates:
+     * - locale (current language)
+     * - flash messages
+     * - connected user object, if any
+     */
     router.use((req, res, next) => {
         app.locals.moment.locale(req.getLocale());
 
@@ -61,23 +123,24 @@ module.exports = function (app) {
         next();
     });
 
+    // For EVERY route of the app, first check if a user is logged in or not
     app.all("*", Controller.isUserLogged);
 
+    // Login/logout routes
     router.get('/login', Controller.needLogin);
     router.post('/login', passport.authenticate('local', {
         failureRedirect: '/login',
         failureFlash: true
-    }), (req, res) => {
-        res.redirect('/');
-    });
+    }), Controller.postLogin);
     router.get('/logout', Controller.logout);
 
+    // Account recovery routes
     router.get('/recover-account/:recoveryID/:recoveryToken', Controller.$checkRecoveryToken, Controller.recoverAccount);
     router.post('/recover-account/:recoveryID/:recoveryToken', UserForm.changePassword, Controller.$checkRecoveryToken, Controller.recoverAccountChangePassword);
     router.get('/recover-account', Controller.recoverAccountForm);
     router.post('/recover-account', Controller.recoverAccountSend);
 
-    /* GET home page. */
+    // Home page
     router.get('/', Controller.index);
 
     // Javascript locales
@@ -94,10 +157,9 @@ module.exports = function (app) {
     router.get('/change-password', Controller.changePasswordForm);
     router.post('/change-password', UserForm.changePassword, Controller.changePassword);
 
+    // App wise API endpoint to compute nearest predefined color
+    // Should be moved in a tool or API router instead
     router.get('/api/color/nearest-predefined/:code', Controller.nearestPredefinedColor);
 
     app.use('/', router);
-
-    // Clean expired token account recovery, every day at 23:23
-    schedule.scheduleJob('23 23 * * *', () => User.CleanExpiredRecovery());
 };